script: Chain up `scrollIntoView()` scrolling to parent `<iframe>`s
Calling `scrollIntoView()` on an element within an `<iframe>` will now
scroll scrolling boxes from the parent document(s), as long as they have
the same origin.
