+++ tests/wpt/meta/css/cssom-view/scrollIntoView-fixed.html.ini
@@ -1,6 +1,3 @@
 [scrollIntoView-fixed.html]
-  [[Box B\] scrollIntoView from unscrollable position:fixed in iframe]
-    expected: FAIL
-
   [[Box D\] scrollIntoView from scrollable position:fixed in iframe]
     expected: FAIL
