+++ tests/wpt/tests/css/cssom-view/support/scrollIntoView-iframes-child.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<meta charset="utf-8">
+<style>
+body {
+  margin: 0;
+  padding-top: 1000px;
+}
+#target {
+  height: 100px;
+  width: 100px;
+  background: green;
+}
+</style>
+<div id="target"></div>
+<script>
+target.scrollIntoView({block: "center"});
+
+addEventListener("message", event => {
+  if (event.data === "scrollY") {
+    event.source.postMessage(scrollY, "*");
+  }
+});
+</script>
